Improve TopologicalSort error message in CommandBatchPreparer
